all the commands

+++ Sprites/SpecialVector.cpp
@@ -1,6 +1,6 @@
 #include "SpecialVector.h"
 
-int SpecialVector::getLength() const {
+unsigned long SpecialVector::getLength() const {
     return this->length;
 }
 
@@ -8,6 +8,6 @@ vector<pair<string, vector<double>>> &SpecialVector::getProperties() {
     return this->properties;
 }
 
-void SpecialVector::setLength(int newLength) {
+void SpecialVector::setLength(unsigned long newLength) {
     SpecialVector::length = newLength;
 }
